Basic Scrubbing and Visualizer
Added the ability to scrub through a music file that is currently
playing. Click on the purple and white bar to scrub.

Once the current track is done, the next one in the list plays. If it
gets to the end of the list, it starts over.

Model is broken. Has something to do with model view projection
variable, although I'm not exactly sure what I'm doing wrong since there
are so many objects within that one object. I don't think it was working
even before, so It's not a top level thing. Probably something wrong
with the shader or vertexes.