Attempts to fix Camera
All attempts have failed so far. I've double checked each class for
errors, but I don't see what the problem is. There is definitly a
problem, I just can't figure out what it is or where it is. OpenGL is
not throwing any errors.

Also cannot see if the debug callback function for OpenGL is working as
intended. No callbacks are made. Seems like there would be atleast one
if for no other reason than initilization