Fixes to Projection Matrix and Transfrom Class
Added a transform class which is based on the one created for c++. It
currently uses a mixture of my own math(which I've tested and works) and
some of OpenTK's(which does not seem to work, and is missing some key
functionality).

When programming a transform class in c++ i used my own math. Even
though glm had functionality for non uniform scaling and euler angles I
didn't use it. I went with uniform scaling and implemented my own euler
angle math. OpenTK doesn't seem to have proper support for euler
angles(it uses quaternions instead which is not simple to use) and it
has no way to scale a matrix non-uniformly.

My next task is to evaluate the GLM# library which is a wrapper for glm
in c++ for c#. If can work with OpenTK's rendering functions I should
use that. If not, making conversions from one matrix type to another at
runtime doesn't seem worth it.

Removed the default namespace from the project settings