Animated Textures
Added the start of a UI framework using a canvas. Button support
currently exists.

Added support for animated textures(including animated gifs) by either
passing in a path to an animated .gif or by passing in multiple image
paths. Working on getting the animation speed right. It inherits from
texture, so it can be used in the place of any normal texture. Not sure
of the performance cost, so I won't combine the two classes just yet.

For betting .gif support, I should look at this article:
http://web.archive.org/web/20130820015012/http://madskristensen.net/post/Examine-animated-Gife28099s-in-C.aspx