Added Media Buttons
Added pause, play, next, and previous buttons. Functionality was already
there, but now it's implemented into the UI.

Added a secondary drawing mode for the EQ visualizer. Looks cool, but
also strange. It uses bezier curves to draw a circle that bumps out.
Only works with certain sample sizes as each sample size if a vert on
the circle. Too many and they overlap, too few and it doesn't look like
a circle.