Fine Tuning
Only the tip of the iceburg on trying to fine tune a visualizer
algorithm. I have everything set up to visualize it, I just need to be
feeding in the proper values.

Edited the static Time class to support getting the time since the
application started. Could be done better.

Debug now uses the Time class. These need to becombined into Utility
somehow which means Time cannot be dependant of OpenTK like it is now.