Behaviours, Renderers
Added some code for behaviours and renderers. That meant adding code for
the type 'Bound'. Seems like there is a lot of code that relys on other
code right now. Eventually I'll have all the root stuff done, but that
seems like a long way off. It's really hard to test this stuff, so I can
only assume things are coded properly for now. Unfortunate.