Tag and Layer Manager
Added a tag and layer manager along with proper management of tags and
layers by the game object class.

Converted most if not all functions which returned an IEnumerable to an
array.

Fixed annoying error with mouse going out of the window and messing up
drag widgets for ImGui. The guy who ported ImGui didn't realize that was
an error on his part... You're supposed to pass -1 for the the mouse
position only when it isn't focused. He had it passing -1 when it was no
longer contained in the window...